remove size_t cast in mrb_assert() since ((size_t)n <= SIZE_MAX) is always true

+++ src/dump.c
@@ -89,7 +89,7 @@ get_pool_block_size(mrb_state *mrb, mrb_irep *irep)
       {
         mrb_int len = RSTRING_LEN(str);
         mrb_assert(len >= 0);
-        mrb_assert((size_t)len <= SIZE_MAX);
+        mrb_assert(len <= SIZE_MAX);
         size += (size_t)len;
       }
       break;
@@ -99,7 +99,7 @@ get_pool_block_size(mrb_state *mrb, mrb_irep *irep)
         int len;
         len = mrb_float_to_str(buf, mrb_float(irep->pool[pool_no]));
         mrb_assert(len >= 0);
-        mrb_assert((size_t)len <= SIZE_MAX);
+        mrb_assert(len <= SIZE_MAX);
         size += (size_t)len;
       }
       break;
@@ -108,7 +108,7 @@ get_pool_block_size(mrb_state *mrb, mrb_irep *irep)
       {
         mrb_int len = RSTRING_LEN(irep->pool[pool_no]);
         mrb_assert(len >= 0);
-        mrb_assert((size_t)len <= SIZE_MAX);
+        mrb_assert(len <= SIZE_MAX);
         size += (size_t)len;
       }
       break;
@@ -419,7 +419,7 @@ write_lineno_record_1(mrb_state *mrb, mrb_irep *irep, uint8_t* bin)
 
   uint32_to_bin((uint32_t)diff, bin); /* record size */
 
-  mrb_assert((size_t)diff <= SIZE_MAX);
+  mrb_assert(diff <= SIZE_MAX);
   return (size_t)diff;
 }
 
@@ -599,7 +599,7 @@ write_debug_record_1(mrb_state *mrb, mrb_irep *irep, uint8_t *bin, mrb_sym const
   mrb_assert(ret <= UINT32_MAX);
   uint32_to_bin(ret, bin);
 
-  mrb_assert((size_t)ret <= SIZE_MAX);
+  mrb_assert(ret <= SIZE_MAX);
   return (size_t)ret;
 }
